+++ lib/repository/ignore.dart
@@ -1,21 +1,17 @@
 import 'dart:io';
 
 import 'package:balo/repository/repository.dart';
-import 'package:balo/variables.dart';
+import 'package:balo/utils/utils.dart';
+import 'package:balo/utils/variables.dart';
 import 'package:path/path.dart';
 
 class Ignore {
   final Repository repository;
 
-  File get file => File(
-        join(
-          repository.directory.parent.path,
-          repositoryIgnoreFileName,
-        ),
-      );
-
   Ignore(this.repository);
+}
 
+extension IgnoreActions on Ignore {
   Future<void> createIgnoreFile({
     Function()? onAlreadyExists,
     Function()? onSuccessfullyCreated,
@@ -28,12 +24,12 @@ class Ignore {
         return;
       }
 
-      if (file.existsSync()) {
+      if (ignoreFile.existsSync()) {
         onAlreadyExists?.call();
         return;
       }
 
-      await file.create(recursive: true, exclusive: true);
+      await ignoreFile.create(recursive: true, exclusive: true);
       onSuccessfullyCreated?.call();
     } on FileSystemException catch (e, trace) {
       onFileSystemException?.call(e);
@@ -52,15 +48,84 @@ class Ignore {
         return;
       }
 
-      if (!file.existsSync()) {
+      if (!ignoreFile.existsSync()) {
         onDoesntExists?.call();
         return;
       }
 
-      await file.delete(recursive: true);
+      await ignoreFile.delete(recursive: true);
       onSuccessfullyDeleted?.call();
     } on FileSystemException catch (e, trace) {
       onFileSystemException?.call(e);
     }
   }
+
+  Future<void> addIgnore({
+    required String pattern,
+    Function()? onAlreadyPresent,
+    Function()? onAdded,
+    Function(FileSystemException)? onFileSystemException,
+  }) async {
+    try {
+      List<String> ignores = ignoreFile.readAsLinesSync();
+      if (ignores.contains(pattern)) {
+        onAlreadyPresent?.call();
+        return;
+      }
+
+      ignores.add(pattern);
+
+      ignoreFile.writeAsStringSync(
+        ignores.join(Platform.lineTerminator),
+        flush: true,
+      );
+
+      onAdded?.call();
+    } on FileSystemException catch (e, trace) {
+      onFileSystemException?.call(e);
+    }
+  }
+
+  Future<void> removeIgnore({
+    required String pattern,
+    Function()? onNotFoundPresent,
+    Function()? onRemoved,
+    Function(FileSystemException)? onFileSystemException,
+  }) async {
+    try {
+      List<String> ignores = ignoreFile.readAsLinesSync();
+      if (!ignores.contains(pattern)) {
+        onNotFoundPresent?.call();
+        return;
+      }
+
+      ignores.remove(pattern);
+
+      ignoreFile.writeAsStringSync(
+        ignores.join(Platform.lineTerminator),
+        flush: true,
+      );
+
+      onRemoved?.call();
+    } on FileSystemException catch (e, trace) {
+      onFileSystemException?.call(e);
+    }
+  }
+}
+
+extension IgnoreCommons on Ignore {
+  File get ignoreFile => File(
+        join(
+          repository.repositoryDirectory.path,
+          repositoryIgnoreFileName,
+        ),
+      );
+
+  List<String> get patternsToIgnore => ignoreFile.existsSync()
+      ? ignoreFile
+          .readAsLinesSync()
+          .where((p) => p.isNotEmpty && !p.startsWith("#"))
+          .map((p) => convertPatternToRegExp(p))
+          .toList()
+      : [];
 }