+++ lib/utils/variables.dart
@@ -1,0 +1,10 @@
+const repositoryFolderName = ".balo";
+const repositoryIgnoreFileName = ".baloignore";
+const branchFolderName = "branches";
+const stateFileName = "state.json";
+const defaultBranch = "default";
+const branchStage = "stage.json";
+
+const stagedAtKey = "staged_at";
+const filePathsKey = "file_paths";
+const currentBranchKey = "current_branch";