+++ lib/variables.dart
@@ -1,5 +1,0 @@
-const repositoryFolderName = ".balo";
-const repositoryIgnoreFileName = ".baloignore";
-const branchFolderName = ".branches";
-const stateFileName = ".state";
-const defaultBranch = "default";